pyproject.toml
[project]
name = "hexlet-flask-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
]
